@@ -1,7 +1,7 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { FormGroup, FormControl, FormBuilder, Validators, AbstractControl } from '@angular/forms';
 import { Subscription } from 'rxjs/Subscription';
-import 'rxjs/add/operator/debounceTime';
+import { debounceTime } from 'rxjs/operators';
 
 import { User } from './../../models/user';
 import { CustomValidators } from './../../validators';
@@ -17,7 +17,7 @@ export class SignupReactiveFormComponent implements OnInit, OnDestroy {
   userForm: FormGroup;
   emailMessage: string;
 
-  private sub: Subscription[] = [];
+  private sub: Subscription;
   private validationMessages = {
     required: 'Please enter your email address.',
     pattern: 'Please enter a valid email address.'
@@ -33,7 +33,7 @@ export class SignupReactiveFormComponent implements OnInit, OnDestroy {
   }
 
   ngOnDestroy() {
-    this.sub.forEach(sub => sub.unsubscribe());
+    this.sub.unsubscribe();
   }
 
   save() {
@@ -142,20 +142,17 @@ export class SignupReactiveFormComponent implements OnInit, OnDestroy {
   }
 
   private watchValueChanges() {
-    const sub1 = this.userForm.get('notification').valueChanges
-      // .subscribe(value => console.log(value));
+    this.sub = this.userForm.get('notification').valueChanges
       .subscribe(value => this.setNotification(value));
-    this.sub.push(sub1);
 
     const emailControl = this.userForm.get('emailGroup.email');
-    const sub2 = emailControl.valueChanges
-      .debounceTime(1000)
+    const sub = emailControl.valueChanges
+      .pipe(
+        debounceTime(1000)
+      )
       .subscribe(value => this.setMessage(emailControl));
-    this.sub.push(sub2);
-
+    this.sub.add(sub);
   }
-
-
 }
 
 
