@@ -1,25 +1,24 @@
 import { Directive, Input } from '@angular/core';
 import { Validator, AbstractControl, NG_VALIDATORS } from '@angular/forms';
 
+import { checkServiceLevel } from './custom.validators';
+
 @Directive({
   selector: '[appServiceLevelValidator]',
-  providers: [{
+  providers: [
+    {
       provide: NG_VALIDATORS,
       useExisting: ServiceLevelDirective,
       multi: true
-  }]
+    }
+  ]
 })
 export class ServiceLevelDirective implements Validator {
-  @Input() rMin = 1;
-  @Input() rMax = 3;
+  @Input() rMin: number;
+  @Input() rMax: number;
 
   validate(c: AbstractControl): { [key: string]: boolean } | null {
-    console.log(c.value);
-    if (c.value !== undefined && (Number.isNaN(c.value) || c.value < this.rMin || c.value > this.rMax)) {
-      return {
-        'serviceLevel': true
-      };
-    }
-    return null;
+    console.log(this.rMax);
+    return checkServiceLevel(c, this.rMin, this.rMax);
   }
 }
